requirements.txt
# Core dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0

# ML dependencies
xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2

# Visualization (optional - for advanced GUI)
plotly==5.18.0

# Development
jupyter==1.0.0